Tracker: Consider 307 status code
307 - Temporary redirect

Say if i am using an application that has a home URL of https://application.domain.name and when you go to that page it checks cookies and if login is not there then send you yo https://application.domain.name/login page and once login is succesful it will go back to the old URL https://application.domain.name .

We can still consider this. Atleast for my usecase that is what i need as of now

Headers can be found by doing curl request with `curl -IL https://app.name` and check the headers and response code.